Implement charts using composed API
